// backend/src/db/schemas/secret-import-versions.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const SecretImportVersionsSchema = z.object({
  id: z.string().uuid(),
  version: z.number().nullable().optional(),
  importPath: z.string(),
  position: z.number(),
  isReplication: z.boolean().nullable().optional(),
  isReserved: z.boolean().nullable().optional(),
  importEnv: z.string().uuid(),
  importId: z.string().uuid(),
  createdAt: z.date(),
  updatedAt: z.date()
});

export type TSecretImportVersions = z.infer<typeof SecretImportVersionsSchema>;
export type TSecretImportVersionsInsert = Omit<z.input<typeof SecretImportVersionsSchema>, TImmutableDBKeys>;
export type TSecretImportVersionsUpdate = Partial<Omit<z.input<typeof SecretImportVersionsSchema>, TImmutableDBKeys>>;